Memory usage is based on the peak & real allocation
The memory usage was previously based on the system memory. By system, PHP means the VM. So first, we should use the real allocated memory. Second, we use `memory_get_peak_usage` instead of `memory_get_usage` because the GC (Garbage Collector) can run and free the memory. Thus, a test case can report 0 byte while it has allocated 1Mb for instance. The peak is what is interesting for us, and the allocated peak.